After agent close check that no further log messages are created
Motivation
----------
A data race is occuring in testing between logging in pipelineclient
and reading the number of log messages on log count testing.

Changes
-------
Move where we log that we're shutting down the pipelineclient.
Add a check after agent close that the number of log messages is
the same before and after a 1s sleep.

Change-Id: Ia28862a68d04831af2cc98b347d08e5aecd505b8
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125349
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>