Create a tracer component
Motivation
----------
The tracer touches many parts of the system and we are currently
passing components a lot of properties to allow them to create
new op traces. It'd be better if we used a tracer component that
knows these properties and then we just pass the tracer component
to other components.

Changes
-------
Create a tracer component.
Pass the tracer component to other components.
Remove CreateOpTrace from http component and instead pass the
tracer component directly to components like n1ql and analytics.

Change-Id: I74c6616f63a9529c8430115c10febcfe5c2592d5
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125388
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>