Create a stats component
Motivation
----------
Stats needs to be move off of agent but it doesn't really fit
in the crud component. Stats isn't really a crud operation and
it also has different requirements to the crud calls. The stats
call needs to be able to hook directly into the kv muxer in order
to be able to send requests directly to pipelines.

Changes
-------
Create a stats component.
Make the StatsEx call hook into the new component.

Change-Id: I8eaa2b69baebd960f039e1a29e62437c221da34e
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125389
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>