If ssl is enabled then use https for http endpoints
Motivation
----------
When we receive a cluster map we pull out and create a list of
endpoints. If ssl is enabled we make these endpoints use the
ssl ports however we also use the http scheme intead of https.
For quite a while now golang has supported automatically handling
setting up http2 when the http transport is created. This means that
we don't need to explicitly call http2.ConfigureTransport.

Changes
-------
If ssl is enabled then use https for endpoints.
Remove the http2 explicit setup.
Remove x/net from our dependencies.

Change-Id: I4c0f0013efe3cedc2bb01cc83c3228c5214309e5
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125391
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>