Refactor component/manager names for consistency
Motivation
----------
We have a mix of components and managers with no clear distinction
between them. They should all be components. The names of them on
agent should also be better.

Changes
-------
Changed all managers to be components.
Renamed properties on agent to drop the component/manager suffix
where it makes sense to do so.

Change-Id: I0d14bc115d588be9b5cf0e9db938e2bbec07d1e5
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125396
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>