GOCBC-838: If ssl is enabled then use https for http endpoints
Motivation
----------
When we receive a cluster map we pull out and create a list of
endpoints. If ssl is enabled we make these endpoints use the
ssl ports however we also use the http scheme intead of https.

Changes
-------
If ssl is enabled then use https for http based endpoints.

Change-Id: I17f9eaf01bf2824dc3bd1e64a388ae99631e5c53
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/124580
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>