Wire in memd package
Motivation
----------
We have a new memd package which improves reusability of the
code for reading from/writing to the network. It also nicely
encapsulates all of the memd connection specifics. We should
use it.

Changes
-------
Remove all constants from gocbcore package that exist in memd
package.
Remove memdTCPConn and replace with a wrapper for the new memd.Conn.
Update memdqrequest to use memd.Packet rather than memdPacket.
Fix all of the references that doing the above breaks.

Change-Id: If62aa88afe5caa453a7e83f49668652c5dfd365b
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125398
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>