Update pipeline iterator to use a callback
Motivation
----------
Pipeline iterator should use a callback and be used everywhere
that pipeline iteration takes place rather than allow direct
use of the client mux state.

Changes
-------
Change pipeline iterator to be more of a partial snapshot and
use a callback.
Fix a bug in the cccp round robin logic that was causing it to
picking pipelines incorrectly.

Change-Id: I1e59b0cf187015f5d6c835594528b40267cee485
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125400
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>