Update WaitUntilReady
Motivation
----------
Rather than just waiting until we have a config the WaitUntilReady
operation should ensure that the underlying connections are in
the state specified in the options. For now we are only concerned
with KV connections.

Changes
-------
Remove the WaitUntilConfig operation.
Remove the WaitUntilComponent.
Add WaitUntilReady operation.
Add a WaitUntilReady call to DiagnosticsComponents which polls
pipeline clients for their current connection state.

Change-Id: I023f9496882510a9aed8b347cc1335dd9ed39855
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125403
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>