Don't try cccp polling if a non-default http port is used
Motivation
----------
If the user provides an address with a non-default http port
then we cannot connect to the cluster. We start the cccp poller
which sees no pipelines and loops back round to try again, and
again, and again... As the cccp looper cannot get a config and
does not fallback it means that we never setup any memd connections.

Changes
-------
If a non default http port is used then don't even attempt cccp
polling, skip straight to http polling.

Change-Id: Ia3aceb0ba79588a5c0fb05b920e1a3f2823e1ddd
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125406
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>