GOCBC-853: Return empty certificate if authenticator returns nil
Motivation
----------
When the couchbases protocol is used (and therefore ssl enabled)
alongside PasswordAuthenticator and a root CA then we cause a panic
within the TLS library. This is because we always provide a
GetCertificate callback on our tls Config and if provided then the
library expects that to return a certificate or an error but we
return both as nil.

Changes
-------
Within the GetCertificate call, If the authenticator returns a nil
certificate and a nil error then actually return a blank certificate.

Change-Id: If57503d1146c74496e4228b99234e8d258d8ed65
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125039
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>