Make HTTP service pseudo-async
Motivation
----------
For consistently with kv ops the http operations should be
asynchonous, taking a callback and returning a PendingOp.

Changes
-------
Update HTTP service ops to take a callback and return a PendingOp.
Within each operation wrap the internals in a goroutine.
Expose cancel via httpRequest via the use of context.CancelFunc.

Change-Id: Ifcbe6b2582209f481b0791225cecfb4a64563a17
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125415
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>