Return an error if kv operation performed over GCCCP connection
Motivation
----------
KV operations cannot be performed against a connection that is
using GCCCP as there is no bucket selected. We should fail fast
these operations with a descriptive error.

Changes
-------
If we attempt to route a KV request whilst the bucket type is none
then return a new ErrGCCCPInUse error.

Change-Id: Ic48ad9595dc4bb5d33e0cf0e43c05eef33d80d99
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125420
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>