GOCBC-855: Create a ConfigSnapshot interface
Motivation
----------
We removed functions like VbucketToServer but some teams like FTS
use these functions so we need to add them back in. We should add
them in a way so that several calls can be guaranteed to be
against the same config version.

Changes
-------
Add a ConfigSnapshot interface and an agent level function to get
one.
Update pipelineSnapshot to be internalConfigSnapshot and make it
satisfy the interface so that we can use that.

Change-Id: I6bdeee07eba7f6e04a5fd6495a05cbc0afcfa26d
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125679
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>