Update n1ql and cbas tests
Motivation
----------
The n1ql and cbas tests are failing on our nightly CI. This is
because of various reasons including trying to run cbas against
server versions that don't support it and the queries in the
cancel tests completing before they have chance to be cancelled.

Changes
-------
Update the server versions that cbas is supported against ensure
that it's 6.0+.
Update the cbas tests to ensure that they are checking for the cbas
feature support.
Add a round tripper which supports a delay before fowarding to the
agent round tripper and use that in cancellation tests.

Change-Id: Ibbedeff46cc1fa496b3aefa0b0bc0a64416b2763
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125701
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>