GOCBC-857: Don't send requests until we've seen a cluster config
Motivation
----------
If a user makes a request before we've seen a valid cluster config
then we will send it as best as we can which can mean that we send
to the wrong node. This can lead to a not my vbucket triggering
and returning a config. We can then race between the NMV and config
 polling trying to apply the first config.

Changes
-------
If we receive a request before we've seen a valid cluster config
then write it to the dead pipeline so that it will be requeued
correctly when we get a config.

Change-Id: I9f87e04e659223dbd296e9b9229a0106cd5a705b
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125892
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>