GOCBC-859: Check if channel is nil calling poller controller Done.
Motivation
----------
When we call agent close from gocb we sometimes see it hang forever.
This is down to a race condition between poller controller Done and
Start. As start is called in a goroutine we seem to get into a
situation where we can end up sitting on the Done signal before
Start has assigned an activeController which leads to us
sitting on a nil signal forever.

Changes
-------
Check if the underlying poller controller done signal is nil
before we wait on it within agent.
Ensure that we don't start a looper if the poller controller has been stopped.

Change-Id: I71871de83e51761f1372011e438af16c0f83e383
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125992
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>