GOCBC-685: Added support for dynamic Root CA fetching.
Motivation
----------
In order to support some specific use-cases where the root
certificates for the server are changing, we should support
using a provider for fetching root certificates on-demand.

Changes
-------
Modified the root-ca code such that it uses a provider method
and invokes it each time it needs to get the list of root
certificates for connecting to a node.

Change-Id: Ifaf9ba4ee9f86f2e93b96fc2e09ee0543730732d
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/126165
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Charles Dixon <chvckd@gmail.com>