Close the old deadpipeline on collection support updated
Motivation
----------
When running against a cluster with collections support we end up
logging that an op queue has been drained before closing. This
is because we're trying to drain the dead pipeline and it hasn't
been closed.

Changes
-------
On collection support updated close the dead pipeline along with
other pipelines.

Change-Id: I1fa7dbb451ead045a63366fe721b77e2798276b3
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/126166
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>