GOCBC-797: Add support for running tests with cert auth
Motivation
----------
We should add functionality to the test suite to be able to run
tests using client cert auth so that we can test that functionality
works correctly.

Changes
-------
Add a new flag to tests.
Wire up the flag to read certs from a directory and then set the
correct authentication correspondingly.

Change-Id: Ibd4e9613f36281007c54fcd0933d495ba0836e42
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/126721
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>