GOCBC-874: Always remove reqs from waitingIn if status != success
Motivation
----------
If a persistent request hit a not my vbucket then on the
subsequent resend of the request sending will fail with a canceled
error. This is because the request waitingIn pointer is still
set, even though we have actually removed the request from the
opmap.

Changes
-------
Update resolve request so that if the response status isn't success
then we always set the request waitingIn pointer to nil.

Change-Id: I7dbccf90f7d663be3a7598784d87462adf9dd60c
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/126834
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>