GOCBC-863: Tidy up kvmux OnNewRouteConfig
Motivation
----------
The kvmux OnNewRouteConfig function is a bit confusing and
difficult to follow. We should rename things and tidy up to make
it easier to follow.

Changes
-------
Update OnNewRouteConfig so that mux states are renamed to states
rather than mux's and refactor out some code blocks into methods
to make it more clear what the code is supposed to be doing.

Change-Id: I083548fdfe0ea12a2a662c34baff70b2c40b6226
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/127408
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>