GOCBC-882: Don't shutdown agent on invalid route data
Motivation
----------
When we create a route data which the config manager deeems as
invalid then we close down the agent. Instead of doing that we
should just skip the update for that config.

Changes
--------
When we create an invalid route data then skip the update instead
of shutting down the agent.

Change-Id: Ic4be792d20737fc31240f187a853cb4310d1783b
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/127412
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>