GOCBC-886: Operations with unknown collections should run until cancel
Motivation
----------
When the user uses a collection (name) which is unknown by us then
we will lookup the cid against the server before performing the op.
If the server replies that it also doesn't know the collection then
we invalidate the collection name forever. We shouldn't do that,
instead we should follow our normal retry logic and retry the
operation (triggering another cid lookup) until either the cid is
returned and the op can be sent or the operation is cancelled/
timed out.

Changes
-------
Update logic so that if a cid fetch fails for an operation
then the collection name is marked unknown and the operation
retried.
Remove all references to invalidcid.

Change-Id: I8c34d4f5e1770bbc4e7fcd70168d92f01027e4a3
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/127490
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>