GOCBC-824: Don't type assert nil pointer in ConnectionInfo
Motivation
----------
When adding enhancements to timeout errors we introduced a possible
nil pointer where we can timeout a request before it's sent. This
would mean that there is no connection info on the request and so
the pointer is nil. We don't check for that so we panic.

Changes
-------
Add a nil check for request connInfo and if nil then return an
empty memdQRequestConnInfo.

Change-Id: I6c81c19dc6e810fa46db488794f2d96efe5f4e7e
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/128089
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>