GOCBC-894: Set request timer after request queued
Motivation
----------
We have a data race that can occur when a request times out before
(or as) the timeout timer is created. We can also trigger a strange
behaviour where we return an error from an operation and then also
run the callback returning a timeout error. This can lead to upstream
races in consumers like gocb.

Changes
-------
Don't set timer on the request until after the request has been
qeueued and so any errors are already returned by the operation.
Change timer on request to be an atomic value so that if it triggers
whilst returning the timer then we don't race assigning it on request.

Change-Id: Ib2947ace839f3fa54a5c04a3ee7650f3bc8c7bd3
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/129181
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>