GOCBC-903: Don't fast fail ops using collections if collection support unknown
Motivation
----------
When performing operations that use collections we check if the SDK
supports collections and if not then return an error. If the SDK hasn't
yet seen a cluster map then it doesn't know if collections are supported
and takes that as a no and fails the op.

Changes
-------
Add a pending op queue to collection manager.
Add a route config watcher to collection manager.
If collections are enabled but the collection manager hasn't seen
a config then queue requests.
Once collection manager sees a config then drain the queue either
dispatching or failing requests depending on whether collections
are supported.

Change-Id: I78ed3d0bab0274413c32f97debaefe6ea1500f14
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/129268
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>