GOCBC-868: Expose a way to fast fail WaitUntilReady
Motivation
----------
Various server components need a way to immediately fail connections
and get the error, if an error occurs at startup time.

Changes
-------
Add an option to WaitUntilReady allows fast fail.
Add a connect error to pipeline client which is set if getting
a memdclient fails.
Return that error through WaitUntilReady if the new option is set.

Change-Id: Iebd08840165db69646121297300342f4df8741e2
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/131570
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>