GOCBC-906: Don't update enhanced durability support state until first config
Motivation
----------
When performing KV ops using enhanced durability without using
WaitUntilReady it's possible to run into an unsupported features error.
This is because when we setup the initial kv mux state we see that the
fake config doesn't have support so mark it as unsupported until we see
a config that says otherwise. This should remain unknown until we see
an actual config.

Changes
-------
If the config ID indicates that it's a fake config then don't
update enhanced durability support status.

Change-Id: Id3f33db837893c21ec1beeb21a19943d23697706
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/129484
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>