GOCBC-899: Update test MakeDistKeys spin until connected or timeout
Motivation
----------
TestRandomGet randomly fails against the mock. This is happening
because the test can be run before the agent is actually completely
connected. In the MakeDistKeys call we get the kv mux state and use
that to work out how many keys to make so we need to spin on
getting the state until either we have a real config or we timeout.

Changes
-------
Update MakeDistKeys to wait until we've seen a config or we
hit a deadline specified.

Change-Id: I9ea28940fb7f211fa24d383571d40e11a3426bed
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/129522
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>