GOCBC-908: Call upstream callbacks if bootstrap ops timeout
Motivation
----------
We're seeing an issue where the agent does not attempt to reconnect
if bootstrap operations timeout (e.g. if the memcached service is
stopped on the server). We do timeout but then never call upstream
callbacks to allow the memdclient to continue and so it never returns
to the dialler component.

Changes
-------
Update doBootstrapRequest logic so that timeouts trigger the callbacks
correctly.

Change-Id: Ic58af50acec469e0a33391a5ead6ad60baf78bd0
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/129589
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>