GOCBC-914: Add cid to ops queued before server collection status known
Motivation
----------
If WaitUntilReady is not used and an operation is performing using a
named collection then the operation enters a queue in the collections
manager. Once the server is known to support collections we send the
request without adding a collection id..

Changes
-------
Once the server is known to support collections then internally
redispatch the request so that the cid fetch is performed and the
cid added to the request.

Change-Id: Ia4411bfff37c392c02144d0b7cb346d4ac6a89ed
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/129693
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>