GOCBC-913: If no Ping services specified don't error for missing services
Motivation
----------
At the moment if a user does a Ping specifying no services against
a cluster which does not have a service such as analytics then we
will include that service in the results as unavailable. We should
instead only ping services that are in the cluster config if no
services are specified.

Changes
-------
Only return errors for missing services if no service types are
specified.
Add a private field to PingOptions so that clusterAgent can instruct
the Ping command to not return errors for missing services.
Add the management service to ping.

Change-Id: I042ccc9da64e7902fb8f5027857fbe45e2344690
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/129833
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>