GOCBC-825: Selectively fail requests after a failed write
Motivation
----------
When performing a write to a memcached connection in the client, we
should be inspecting the write error to determine whether it was an
unambiguous hard error (guarentees that the remote end never processed
the request). Or if it was a ambiguous kind of timeout (not sure whether
 the remote end processed the request).

Changes
-------
Add a WriteError to the memd package and return it for any errors
that occur on the Write call, wrapping the underlying error.
Add a corresponding retryreason and run it through the orchestrator
if the WriteError is returned from the WritePacket call.
Don't cancel the request in memdclient after a failed write, instead
leave that to the memdpipelineclient to decide depending on result
of call to retry orchestrator.

Change-Id: I654f2824816053c0f8ccb2ae6594102b9fd2cd4f
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/130001
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>