GOCBC-915: Tidy up memdclient bootstrap logic
Motivation
----------
doBootstrapRequest is a bit convoluted and difficult to follow.
The function should just take a memdqrequest with a callback
already set on it.

Changes
-------
Update doBootstrapRequest to take a memdqrequest and no callback.
Update callers to create a memdqrequest directly handling the
callback themselves.
Update doBootstrapRequest to use memdquest.SetTimer instead of
tracking a timer ourselves.

Change-Id: Id41358fbe912ee174559c57c07563debc15ddd6c
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/130140
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>