GOCBC-921: Update cid refresh logic to never set op queue to nil
Motivation
----------
When we get an error for the cid refresh request we set the cache
op queue to nil. This can cause a panic if we have a different
request come in at the same time and try to get added to the queue.

Changes
-------
Update the logic so that it remove the entire cache from the
collections manager if there's an unknown error.
Add locks around modifications to the op queue in the cid refresh
callback.
Add a check to ensure that the cid state is set back to unknown if
the refresh failed to send.
Add lots more logging.

Change-Id: I3ab14d911bedf336478d73a8a64ce2ac4ef438e9
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/130225
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>