GOCBC-919: Improve linting abilities
Motivation
----------
We've been using the same linting for quite some time now and IDEs
are beginning to flag up issues that our CI isn't. We've moved gocb
to use golangci-lint so we should do the same for gocbcore.

Changes
-------
Replace all existing linting with golangci-lint which covers the
linters we were using and many more.
Pick a sensible set of linters for now, allowing us to add more in
the future.
Make various code fixes to satisfy the linters.

Change-Id: I0d30e4e8db05ec88e4a2006581bbd400270e0ba7
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/130773
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>