GOCBC-925: Use n1ql Err to return any extra errors from the result
Motivation
----------
Sometimes the server can respond to a request with a http 200 status
code and both errors and results. The results array comes before
the errors and as we stream results we have no knowledge of the errors
before returning a result to the user. We need to hijack the Err call
and check if there are any errors in the metadata bytes, if so then
return the errors.

Changes
-------
If there is no stream error on the Err call then attempt to parse
the metadata errors key and return any errors that we find.

Change-Id: I92ec1954e5b3787d9a9974c96d8b3198763056a7
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/131109
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>