GOCBC-928: Update key we use to enable enhanced durability
Motivation
----------
When we receive a cluster config from the server we check the bucket
capabilities to see if it supports enhanced durability. We're
using the incorrect key.

Changes
-------
Update kv mux state to check for durableWrite instead of syncreplication.
Add a few tests to try ensure that this doesn't happen again.

Change-Id: Ifef1115dd6b06f103b29159ddb560c648257caf8
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/131243
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>