GOCBC-929: Propagate errors to stream handler End
Motivation
----------
If an EOF occurs whilst streams are open then it is expected that
the errors will propagate to the End handler of the stream observer.
At the moment they are sent to the OpenStream callback even if that
callback has already been called.

Changes
-------
Update our DCP handler logic so that we use an atomic to determine
if we've already seen the OpenStream callback and if so then
route errors where there is no OpenStream response to the End
handler.

Change-Id: I7af7debd7c12ed8cee9025aa89fd2c29fb38623d
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/131271
Reviewed-by: James Lee <james.lee@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>