GOCBC-933: Fallback to HTTP polling if select bucket unsupported
Motivation
----------
When using a single bootstrap address which is also a non-kv node
then select bucket will fail. This will cause the CCCP poller's
cccp request to never get sent meaning that the SDK will never
receive a cluster config. When select bucket return ENOENT then
we should fallback to HTTP to try to fetch the cluster config.

Changes
-------
Update CCCP poller so that if the cccp request times out
then we should check the underlying connections for any
ENOENT errors. If found then we should fallback.
Update CCCP poller to return a new error if there are no pipelines,
to allow the http poller to run.
Update agent to always provide the poller controller with a cccp
poller so that it can be used for "fall forward".
Add a config watcher to poller controller and on first config
check if the bucket is couchbase type and if the http poller is in
use, if so then start the cccp poller.

Change-Id: I2d374a481a155a1624568de01abe4ba5a8e5fd6f
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/131941
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>