GOCBC-949: Add collections support to GetRandom
Motivation
----------
The server has now added collections support to the GetRandom
command so we should do too. Tests are now failing because if
collections are enabled then the server expects a cid for the
command.

Changes
-------
Add the usual scaffolding for collection name, scope name and
collection ID options to GetRandom.
Update the memd/conn to correct embed the cid into the extras.

Change-Id: I2eb9ca9f09a140702b177d906b5368fbcd03cc1e
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/132116
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>