GOCBC-868: Add fast fail waituntilready for non-default http bootstrap
Motivation
----------
We added fast failing behaviour for waituntilready to allow users
to get the underlying connection errors as soon as they're hit.
This doesn't work for bootstrapping on http against non-default
ports because doing that doesn't attempt to create any memcached
connections. This means that we need to be able to get the error
from the http poller itself.

Changes
-------
Add the ability to bubble up any errors from the pollers.
In waituntilready if we aren't connected and there are no errors
from the connections then check if the pollers have any non-timeout
errors.

Change-Id: Idabeb251a7cbef156c2f62c803d38cdb7512f3e6
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/132235
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>