GOCBC-955: Don't reorder xattr subdoc ops
Motivation
----------
When sending subdocs we need to reorder the operations so that
Xattrs are sent first, however the order that the xattrs are sent
is important and must be kept the same.

Changes
-------
Remove subdocOpList Append and Prepend and replace with Reorder.
Make Reorder persist the order in which xattrs are ordered when
moving them to the head of the op list.

Change-Id: Ia00d012d1198eb2a3d1b20ab4b976ed897639415
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/132646
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>