GOCBC-720: Add tracing verification to tests
Motivation
----------
We should enhance our tests so that we verify that the correct
tracing is taking places for the operations being performed.

Changes
-------
Add a test tracer implementation and helpers for verifying that
the tracer has received the correct spans during the test.
Add testing for tracing to most relevant tests.

Change-Id: Iacedb3fe846154175878a7c8aa182c1609a4e311
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/133354
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>