GOCBC-868: Add bucket not found retry reason
Motivation
----------
We now have functionality to fail fast from WaitUntilReady and
that behaviour can be controlled via retry reasons. However, we
need to expose a user to differentiate a bucket that cannot be
accessed and one that the user can access but is currently not
usable.

Changes
-------
On SelectBucket convert any doc not found error into a bucket not
found error.
Add a bucket not found retry reason.
Expose the bucket not found retry reason to the retry strategy if
that is the underling reason for connections failing.

Change-Id: I3c5f9f4a2e3ba28ea87d5d67c1ed63fed5cb02f1
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/133300
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>