GOCBC-978: Only allow single request to trigger collection pending
Motivation
----------
It's possible for more than one request to trigger the behaviour to
switch over from unknown to pending state for a given collection/scope.
If two or more requests come in at the same time then they can all
trigger the behaviour at the same time. This means that the get cid
request is sent multiple times and could also lead to some undefined
behaviour.

Changes
-------
Update getting and adding caches to the collection manager so that
the get and add is done within a single lock.
Update GetCollectionID so that it performs an upsert of the cache
within a lock too.

Change-Id: I46b4ae1921d2e244672f210f812a3324882f481f
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/134928
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>