GOCBC-981: Do not change connection string network option
Motivation
-----------
When a user uses the "network" option in the connection string
and uses the value "default" then we change the value to "". We
shouldn't do that as it prevents them from being able to force
the SDK to use the "default" network.

Changes
-------
Update connection string network parsing to not change the value
"default" to "".

Change-Id: Ib3a76773adb1cf6605ff666290ff14a123c82677
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/135072
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>