GOCBC-981: Do not change connection string network option
Motivation
-----------
When a user uses the "network" option in the connection string
and uses the value "default" then we change the value to "". We
shouldn't do that as it prevents them from being able to force
the SDK to use the "default" network.

Changes
-------
Update connection string network parsing to not change the value
"default" to "".

Change-Id: I2682a3d316947036150923dd8b569c92f06c8149
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/135073
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>