GOCBC-984: Fixed DCP backpressure to not block the client.
Motivation
----------
The DCP implementation in the Go SDK intentionally moves all DCP
event handling to another goroutine which is separate from the
IO goroutine to avoid blocking all IO when DCP message
back-pressure is used. A bug in the SDK where the channel which
connects these two goroutines was unbuffered meant that DCP
events could cause a blocked IO goroutine, which prevented a
number of important behaviours such as config refreshes and
further DCP requests (OpenStream) from being made.

Changes
-------
Updated our configuration to allow configuration of the DCP
buffer sizing as well made the dcp queue size based on this.
Fixed the DCP item queue to be sized appropriately.

Change-Id: I122709175920a247213f9c05687e7128c3d379cd
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/135469
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>