GOCBC-982: Do not change connection string network option for DCP
Motivation
-----------
When a user uses the "network" option in the connection string
and uses the value "default" then we change the value to "". We
shouldn't do that as it prevents them from being able to force
the SDK to use the "default" network.

Changes
-------
Update connection string network parsing to not change the value
"default" to "".

Change-Id: I8768d0b90b90e16b6b56fb6bbb66774aee05d320
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/136143
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>