GOCBC-984: Fixed DCP backpressure to not block the client.
Motivation
----------
The DCP implementation in the Go SDK intentionally moves all DCP
event handling to another goroutine which is separate from the
IO goroutine to avoid blocking all IO when DCP message
back-pressure is used. A bug in the SDK where the channel which
connects these two goroutines was unbuffered meant that DCP
events could cause a blocked IO goroutine, which prevented a
number of important behaviours such as config refreshes and
further DCP requests (OpenStream) from being made.

Changes
-------
Updated our configuration to allow configuration of the DCP
buffer sizing as well made the dcp queue size based on this.
Fixed the DCP item queue to be sized appropriately.

Change-Id: Id31b97c64d42623cbe3675fcb96a6e31b8a8e84f
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/136292
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>