GOCBC-987: Use a boolean for collections featurein memd conn
Motivation
----------
The memd subpackage currently has a map of all protocol features which
are enabled. This is used to determine how to encode and decode packets.
We should add a special-case for collections to reduce the overhead of
encoding packets by using a boolean check rather than map lookup and
then boolean check.

Changes
-------
Update memd conn to detect if collections feature is being enabled
and if so set a boolean directly as well as adding to the features
map.
Whenever collections support is checked use the boolean rather than
the map.

Change-Id: I882ae2b37ac77decb020e989d4b5dfa798b15d17
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/136468
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>