GOCBC-991: Don't fail bootstrap if auth error was unsupported mechanism
Motivation
----------
When we perform authentication against the server the first attempt
is a best guess. The server can reply that this is an unsupported
mechanism and we should then compare the mechanisms we have enabled
against what the server supports and try again until that list
is exhausted. We already implement this but the code never gets
reached because we always fail bootstrap on auth error. Instead
of doing this we should check if the mechanism we just tried is
in the list of server supported mechanisms and if so then fail
bootstrap.

Changes
-------
Update authentication in bootstrap so that if we receive an auth
error then we check if the mechanism we just tried is in the list
of server supported mechanisms and if so then fail bootstrap, otherwise
try to find the next mechanism.

Change-Id: I491cd1c20b6a2cbaace46bf0e19cc2e8022d1603
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/136706
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>