GOCBC-998: Add a 'GetAllCollectionManifests' function
Motivation
----------
Collection manifest updates are not atomic across the cluster meaning
it's possible for one or more nodes to be slightly behind for a short
period of time.

The current 'GetCollectionManifest' function always dispatches to the
node containing vBucket 0/replica 0 meaning the caller could perform KV
operations on a scopes/collection which is not ready to take traffic on
another node in the cluster.

Changes
-------
Add a 'GetAllCollectionManifests' function which will fetch the manifest
from each node in the cluster. This will allow users to check if updates
to the manifest have been propagated to all the nodes in the cluster.

Change-Id: Ic59972033bfa744557a2f48531b4db0aa9477155
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/137558
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: James Lee <james.lee@couchbase.com>