GOCBC-1007: Only expose unknown errors to error map handling
Motivation
----------
The servers error map is telling us that durability impossible is a
retryable error. This is incorrect and we should not retry it, this
means that we need handle it SDK side. This should already happen
because we shouldn't expose errors to the error map unless we don't
know the error.

Changes
-------
Update kv mux retry handler to only retry error map errors if we
don't know what they are.

Change-Id: Ib1344ec1d2a3b79e704917a478a087c5578a255e
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/139365
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>