GOCBC-1013: Add new keys and chunks fields to stats result
Motivation
----------
The server response for some stats keys does not contain simple types
per packet. Instead the packets contain complex objects. We need a way
to allow users to access these values.

Changes
-------
Only consider the stream terminated if both key and value are empty
for a packet.
Add StatsKeys and StatsChunks to SingleServerStats.
For each packet that we receive add the raw key to StatsKeys and the
raw value to StatsChunks.

Change-Id: I7d6bc400054828b1f66e2ce987045d2b00872bfb
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/140375
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>