GOCBC-1016: Set diagnostics poller controller to a hard nil if no poller
Motivation
----------
Creating an agent with no bucket and a http address on a non default
port can lead to a pnic within WaitUntilReady if used. This is because
we use an interface for pollerErrorProvider which when set to the agent
pollercontroller pointer gives the interface a type. This means that the nil
check does not fail despite the value being nil.

Changes
-------
Move diagnostics creation into the poller controller nil check so
that we can set the pollerErrorProvider to a hard nil if required.

Change-Id: Iea7a93929ec922e4fbda14eab5a9c25e1a4dba11
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/140632
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>