GOCBC-1008: Convert exists error to cas mismatch for subdoc
Motivation
----------
When applicable we covert full doc DocAlreadyExists errors to
CasMismatch. We need to do the same for subdoc mutations i.e. do
the conversion when the doc level flags are not set to Add.

Changes
-------
Update the translateMemdError to convert exists error to cas mismatch
for mutatein when the operation is not a doc level Add.

Change-Id: I863b6e6c21d272f374b711345be003d55fcdee9d
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/141342
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>