GOCBC-854: Add support for user impersonation
Motivation
----------
Sometimes services within the server need to be able to perform
operations on behalf of other users, while still authenticating
with the server using their own credentials.

Changes
-------
Add a new frame type to the memd packet and wire into WritePacket.
Add an internal option to kv operations to set a user to
impersonate.

Change-Id: Ia9da68c9632fba3f889533785b696fb9fa3864f8
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/141603
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>