GOCBC-1028: Fallback to http polling if error is bucket not found
Motivation
----------
When bootstrapping against a non kv node connecting fails. This
happens because select bucket fails with a key not found and we
convert this to bucket not found. The cccp poller request times
out so it goes looking in the pipelines and finds the bucket not
found error, this isn't a poller fallback error so it keeps trying
and we never manage to connect due to select bucket never succeeding
and us never fetching a cluster map.

Changes
-------
Add bucket not found as a poller fallback error.

Change-Id: I786ba5f90d4d3632315507995e4e4c13cd9374aa
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/141905
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>