GOCBC-1030: Fallback to http config fetches on some select bucket fail
Motivation
----------
When select bucket fails due to an error such as key not found then
we can be certain that CCCP is also going to fail. At present when
this happens we wait for the CCCP request to timeout before looking
into the pipelines to work out why and possibly falling back.
Instead of doing this we should bubble any bootstrap errors up
and fallback to the http looper if applicable.

Changes
-------
Add a bootstrap fail handler to memd dialer.
If the bootstrap error is a fallback error then force a http
poller fallback, if not already running.

Change-Id: I72d7072c399ec3aca120f7944d6c0b105d8b2e13
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/142096
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>