GOCBC-1030: Allow pipeline client to shutdown without waiting for dialler
Motivation
----------
When we shutdown a pipeline client we wait for the client to be
fully shutdown before we return from Close. If the pipeline client
has just begun dialling a new memdclient then this can take some
time due to bootstrap needing to complete first. This can make
takeover on a new config quite slow if there are multiple pipelines
to close and/or pipelines have many clients.

Changes
-------
Add a new cancel dial channel to pipeline client which can be
use to signal to cancel fetching a new memd client.
Add a channel to bootstrap and dial which can be used to cancel.

Change-Id: Ie5ed835e7e6fc08acf1ff5caaaf70eecf958a211
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/142098
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>