GOCBC-1040: Include error causes when doing serialization
Motivation
----------
When we serialize our error types we omit the underlying "inner"
error from the serialized data. We should include this information
in the serialized data.

Changes
-------
Add custom MarshalJSON functions to our error types and convert
the errors to anonymous structs so that we can stringify the
"inner" error before marshalling the error.

Change-Id: I9d5c9c87848938397fe8a333774af94f85c1bb80
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/143819
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>